video: fsl-dcu-fb: fix pixel clock polarity
The pixel clock polarity setting was still wrong: In contrast what
earlier mentioned, the reference manual is correct! But the kernel
display flags are rather somewhat confusing: The flags specify the
edge where the data should be driven by the controller (and hence not
sampled by the display!). We convert the display flag to the display
centric DCU notation when getting the videomode from device tree.