imx6q/dl.dtsi: Bump set point voltages
This commit evenly bumps voltages by 35mV from the minimum voltages
found in IMX6DQCEC.pdf  (ignoring the highest freq as numbers aren't
readily available). This 35mV tolerance follows the assumptions made in
AN4724 + testing on a variety of boards and verifying the value
empirically. 35mV should be enough of an increase to account for loss
through traces, inefficiencies of the LDO internal to the processor etc.

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>